scripts/flying-banana/requirements.txt
google-generativeai>=0.8.1
PyGithub>=2.4.0
